pyproject.toml
[project]
name = "automaton-auditor"
version = "1.0.0"
description = "Deep LangGraph Swarm for Autonomous Governance"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langsmith>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.10.0",
    "docling>=2.0.0",
    "pillow>=10.0.0",
    "gitpython>=3.1.0",
    "aiofiles>=23.0.0",
    "asyncio>=3.4.0",
    "tenacity>=8.0.0",
    "structlog>=24.0.0",
    "rich>=13.0.0",
    "langchain-core>=1.2.15",
    "tree-sitter-python>=0.25.0",
    "pymupdf>=1.27.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "black>=24.0.0", "ruff>=0.3.0", "mypy>=1.8.0"]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = []

[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "langsmith>=0.7.6",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]
